changed type of price in subscription from double to decimal

+++ Backend/Project/Helper/SubscriptionPriceCalculator.cs
@@ -5,21 +5,21 @@ namespace Project.Helper;
 
 public static class SubscriptionPriceCalculator
 {
-    public static double CalculateAmount(double defaultPrice, User user)
+    public static decimal CalculateAmount(decimal defaultPrice, User user)
     {
-        double discount = 0;
+        decimal discount = 0;
 
         switch (user.Status)
         {
-            case Status.Student: discount += 0.20; break;
-            case Status.Elder: discount += 0.10; break;
+            case Status.Student: discount += 0.20m; break;
+            case Status.Elder: discount += 0.10m; break;
         }
 
         switch (user.Country.Trim().ToUpperInvariant())
         {
-            case "POLAND": discount += 0.05; break;
-            case "GERMANY": discount += 0.03; break;
-            case "FRANCE": discount += 0.02; break;
+            case "POLAND": discount += 0.05m; break;
+            case "GERMANY": discount += 0.03m; break;
+            case "FRANCE": discount += 0.02m; break;
         }
 
         return Math.Max(defaultPrice * (1 - discount), 0);
